requirements.txt
# Core dependencies
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0  # PostgreSQL support
aiosqlite>=0.19.0  # SQLite async support

# Vector embeddings and semantic search
sentence-transformers>=2.2.0
numpy>=1.24.0

# LLM integrations
openai>=1.0.0
anthropic>=0.7.0
litellm>=1.0.0

# Graph support (optional)
networkx>=3.0

# Async support
aiohttp>=3.9.0
asyncio>=3.4.3

# Configuration and utilities
pydantic>=2.0.0
python-dotenv>=1.0.0

# Logging and monitoring
structlog>=23.0.0

# Testing (for development)
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-asyncio>=0.21.0

